[v6.1.1] 整理 `zhengcheng-parent` 项目的系统模块 微服务 ums 分页

+++ zhengcheng-common/src/main/java/com/zhengcheng/common/domain/PageQuery.java
@@ -29,12 +29,12 @@ public class PageQuery implements Serializable {
     /**
      * 分页大小
      */
-    private Integer pageSize;
+    private Long pageSize;
 
     /**
      * 当前页数
      */
-    private Integer pageNum;
+    private Long pageNum;
 
     /**
      * 排序列
